feat: 로그인 API 구현 및 JWT/FCM 토큰 처리 로직 추가
- 로그인 시 accessToken, refreshToken 발급 및 DB 저장
- JWT 토큰 생성 시 사용자 ID 기반으로 생성
- 로그인 요청 시 FCM 토큰 전달받아 저장하는 로직 추가
- JWT 관련 클래스 jwt 패키지로 분리
- Swagger 예시값 추가

S12P21A304-74

+++ backend/src/main/java/newstock/domain/user/entity/User.java
@@ -14,7 +14,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId; // 유저 고유 아이디
+    private Integer userId; // 유저 고유 아이디
 
     @Column(unique = true)
     private String email; // 유저 로그인 아이디(이메일)
